removing unneeded stuff
